implement query and connect to input
