+++ src/containers/Test.js
@@ -1,97 +1,45 @@
 import React from 'react';
-import Clarifai from 'clarifai';
-import GenreToDescriptive from '../utils/genre-to-descriptive';
-//
-
-const app = new Clarifai.App({
-	apiKey: '85c245c94f054d909333a497016a9a92',
-});
+import Input from '../components/Input';
+import GetDataFromHasura from '../components/Query';
 
 class Test extends React.Component {
 	constructor(props) {
 		super(props);
-		this.state = { input: '' };
-		this.handleSubmit = this.handleSubmit.bind(this);
-		this.handleChange = this.handleChange.bind(this);
+		this.state = { searchGenre: '' };
+
+		this.upDateSearchGenre = this.upDateSearchGenre.bind(this);
 	}
 
-	handleChange(e) {
-		this.setState({ input: e.target.value });
-	}
-
-	handleSubmit(e) {
-		e.preventDefault();
-		let url = this.state.input;
-		this.recognitionWithClarifai(url);
-		this.setState({ input: '' });
-		//displayVideo(url);
-	}
-
-	recognitionWithClarifai(url) {
-		app.models
-			.predict(Clarifai.GENERAL_MODEL, url, { video: true })
-			.then(response => {
-				let conceptsOfVideo = this.reduceAPIresponseToOneArray(response);
-				console.log(conceptsOfVideo);
-				let matches = this.showMatch(conceptsOfVideo, GenreToDescriptive);
-				//displayMatches(matches);
-				console.log(matches);
-			})
-			.catch();
-	}
-
-	reduceAPIresponseToOneArray(response) {
-		let frames = response.outputs[0].data.frames.map(frame => {
-			return frame.data.concepts;
-		});
-		let conceptsPerFrame = frames.map(frame => {
-			let frameConcepts = [];
-			frame.forEach(concept => frameConcepts.push(concept.name));
-			return frameConcepts;
-		});
-		let conceptsOfVideo = [];
-		conceptsPerFrame.forEach(frame => {
-			frame.forEach(concept => {
-				if (!conceptsOfVideo.includes(concept)) {
-					conceptsOfVideo.push(concept);
-				}
-			});
-		});
-		return conceptsOfVideo;
-	}
-
-	showMatch(conceptsOfVideo, GenreToDescriptive) {
-		let matches = {
-			oriental: 0,
-			jazz: 0,
-			rock: 0,
-			folk: 0,
-			ambient: 0,
-		};
-		conceptsOfVideo.forEach(concept => {
-			GenreToDescriptive.forEach(genre => {
-				if (genre.descriptive.includes(concept)) {
-					matches[genre.name]++;
-				}
-			});
-		});
-		return matches;
+	upDateSearchGenre(genre) {
+		this.setState({ searchGenre: genre });
 	}
 
 	render() {
 		return (
 			<div>
 				<h1>Test page</h1>
-				<form onSubmit={this.handleSubmit}>
-					<input
-						className="link-input"
-						type="text"
-						placeholder="put link to your video here"
-						value={this.state.input}
-						onChange={this.handleChange}
-					/>
-					<input type="submit" value="Submit" />
-				</form>
+				<p>
+					Try with the following links:
+					<br />
+					https://s3.us-east-2.amazonaws.com/freecodecamp-hackaton/videos-for-testing/Dunes1.mp4
+					<br />
+					https://s3.us-east-2.amazonaws.com/freecodecamp-hackaton/videos-for-testing/Istanbul.mp4
+					<br />
+					https://s3.us-east-2.amazonaws.com/freecodecamp-hackaton/videos-for-testing/coffe.mp4
+					<br />
+					https://s3.us-east-2.amazonaws.com/freecodecamp-hackaton/videos-for-testing/traffic.mp4
+					<br />
+				</p>
+
+				<Input upDateSearchGenre={this.upDateSearchGenre} />
+				{this.state.searchGenre ? (
+					<div>
+						<h3>Genre Match: {this.state.searchGenre}</h3>
+						<GetDataFromHasura keyword={this.state.searchGenre} />
+					</div>
+				) : (
+					''
+				)}
 			</div>
 		);
 	}
